# docs/swagger/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  gin.Context:
    properties:
      accepted:
        items:
          type: string
        type: array
      errors:
        type: errorMsgs
      keys:
        additionalProperties: true
        type: object
      params:
        type: Params
      request:
        type: string
      writer:
        type: ResponseWriter
    type: object
  set.SetResponseForm:
    properties:
      body:
        example: '{"data":"your response data"}'
        type: string
      code:
        example: 200
        type: integer
      header:
        example: '{"content-type":"application/json"}'
        type: string
    type: object
info:
  contact: {}
  description: A server using to test
  license: {}
  termsOfService: https://github.com/JiHanHuang/stub
  title: Golang Stub
  version: "1.0"
paths:
  /admin/v1/define/resp:
    delete:
      parameters:
      - default: set_response
        description: 自定义返回名
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
      summary: 删除自定义返回列表
      tags:
      - Admin
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
      summary: 自定义返回列表
      tags:
      - Admin
    post:
      parameters:
      - description: 设自定义返回结构
        in: body
        name: setResponse
        schema:
          $ref: '#/definitions/set.SetResponseForm'
      - default: set_response
        description: 自定义返回名
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 设置自定义返回
      tags:
      - Admin
  /api/v1/check/pass:
    get:
      parameters:
      - description: 匹配自定义返回(可选)
        in: query
        name: "yes"
        type: string
      - description: 不匹配自定义返回(可选)
        in: query
        name: "no"
        type: string
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 校验账号密码，若密码为账号的base64，则返回yes，否则返回no。支持自定义返回 [get|post]
      tags:
      - API
  /api/v1/data/delay:
    get:
      parameters:
      - description: 延时时长(默认5s)
        in: query
        name: delay
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 延时返回 [get|post]
      tags:
      - API
  /api/v1/data/normal:
    get:
      parameters:
      - description: 数据量(k)默认0k
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取一定量的数据 [get|post]
      tags:
      - API
  /api/v1/data/p:
    get:
      parameters:
      - description: 数据量(0,1,10)默认0
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取一定量的数据(并发) [get|post]
      tags:
      - API
  /api/v1/define/resp:
    get:
      parameters:
      - default: set_response
        description: 自定义返回(可选)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取自定义返回数据 [get|post]
      tags:
      - API
  /api/v1/file/del:
    delete:
      parameters:
      - description: file name
        in: query
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 删除上传文件 [delete]
      tags:
      - API
  /api/v1/file/download:
    get:
      parameters:
      - description: file name
        in: query
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.Context'
      summary: 下载文件(需要文件上传接口上传) [get]
      tags:
      - API
  /api/v1/file/download2:
    get:
      parameters:
      - description: file name
        in: query
        name: filename
        required: true
        type: string
      summary: 下载文件(不可靠)(需要文件上传接口上传) [get]
      tags:
      - API
  /api/v1/file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 上传文件 [post]
      tags:
      - API
  /api/v1/fingerprint:
    post:
      parameters:
      - description: appkey
        in: query
        name: app_key
        required: true
        type: string
      - default: '{"app_id":"xxxxx",...}'
        description: data
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取数据 [post]
      tags:
      - API
  /api/v1/show/base:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 请求该接口，打印出请求相关信息，更友好显示 [get|post]
      tags:
      - API
  /api/v1/show/url:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 请求该接口，打印出请求相关信息 [get|post]
      tags:
      - API
swagger: "2.0"